+++ src/plonemeeting/restapi/deserializer/atcontent.py
@@ -1,9 +1,8 @@
 # -*- coding: utf-8 -*-
 
 from plone import api
-from plone.restapi.deserializer.atcontent import (
-    DeserializeFromJson as ATDeserializeFromJson,
-)
+from plone.restapi.deserializer import json_body
+from plone.restapi.deserializer.atcontent import DeserializeFromJson as ATDeserializeFromJson
 from plone.restapi.interfaces import IDeserializeFromJson
 from Products.PloneMeeting.interfaces import IATMeetingContent
 from zope.component import adapter
@@ -32,3 +31,13 @@ class DeserializeFromJson(ATDeserializeFromJson):
         return super(DeserializeFromJson, self).__call__(
             validate_all=validate_all, data=data, create=create
         )
+
+    def validate(self):
+        """Handle "ignore_validation_for" that will specifically
+           let ignore validation errors for given fields."""
+        errors = super(DeserializeFromJson, self).validate()
+        data = json_body(self.request)
+        ignore_validation_for = data.get("ignore_validation_for", [])
+        for field_name in ignore_validation_for:
+            errors.pop(field_name, None)
+        return errors